backend/requirements.txt
# Core Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.0

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Web Scraping
beautifulsoup4==4.12.2
selenium==4.16.0
scrapy==2.11.0
playwright==1.40.0
requests==2.31.0
lxml==4.9.3

# AI/ML
google-generativeai==0.3.2
transformers==4.36.2
spacy==3.7.2
nltk==3.8.1

# Email
Flask-Mail==0.9.1
sendgrid==6.11.0
mailgun==0.1.1

# Task Queue
celery==5.3.4
redis==5.0.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
validators==0.22.0
pdfplumber==0.10.3
PyPDF2==3.0.1
Pillow==10.1.0

# Security
bcrypt==4.1.2
cryptography==41.0.7

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==21.0.0

# Development
black==23.12.1
flake8==6.1.0
ipython==8.18.1

# Web Server
gunicorn==21.2.0
gevent==23.9.1

# Monitoring
sentry-sdk==1.39.2
